Improve Jenkins QoS and usability
We need to balance resources on user machiens with Jenkins hunger for resources. We
settled on 1.5 CPU for that to make sure there was enough space for all other
resources to be scheduled.

In addition, we are intentionally providing, and matching, the requests and limits
for both CPU and memory. This will implicitly set the QoS to guaranteed in k8s and
hopefully reduce any unnecessary OOM killings[1].

[1] https://kubernetes.io/docs/tasks/configure-pod-container/quality-service-pod/#create-a-pod-that-gets-assigned-a-qos-class-of-guaranteed