FIX: Run work-creator tests on m1 macs
The script we are using has a default k8s version of 1.19.2. This version does
not have an available kubebuilder-tools for m1 mac[1]. By using the
`ENVTEST_K8S_VERSION` we can override to an m1 friendly version.

In addition, used the explicit gingko version to run the tests just
as we did in the main kratix tests[2].

[1] https://storage.googleapis.com/kubebuilder-tools/
[2] https://github.com/syntasso/kratix/pull/7/files