chore: use jdk17 on jenkins instace
as jdk17 boots up a bit faster, allowing for less flakes on CI

However, the location of `tini` has moved and the previous location (on
/sbin/tini) has a 2 minutes sleep built-in to force people to use the
new location (on /usr/bin/tini).

The `command` being used in the container is exactly the same as the
jenkins/jenkins:2.332.2-lts image, but uses the new `tini` location.

[#183033726]