chore: rename script
the defunct "prepare-platform-cluster-as-worker" script is going to be
incorporated into the workshop and the previous name was a mouthful.

renamed it to "register-worker" for simplicity.

on top of that, we also want the labelling of the platform cluster to
occur as a separate step; therefore the script no longer labels the
cluster, unless `--with-label` is provided

furthermore, steps were taken to make the script more generic