fix: remove stale configmap after promise install
when the promise contains a promise.configure workflow, kratix creates a
configmap with the promise destinationSelectors.

the configmap is used by the work-creator to, in combination with the
workflow destinationSelectors, determine where the workload should go

the configmap is potentially out-of-date after the workflow execution.

to avoid confusion, this commit removes the configmap after a successful
completion of the workflow.

note it will be recreated with the correct information if a resource
request gets created in the `kratix-platform-system` namespace, or if
the promise is updated.

[finishes #186324431]