fix: use gh/release job in workflow
In trying to login, the previous commit attempted to use an orb
job as a command which failed.

In addition, on review it showed that all the files we were
copying to `dist` were the same as all the files in `distribution`
so we optimised a bit by removing that copy.

Co-authored-by: Chris Hedley <chris@syntasso.io>