feat: attempt to preserve version id on failures
there are multiple points in the codebase where the workplacement object
gets updated, which can cause the object to be out of date at time of
updating the status

ideally i'd like to untangle the updates in such a way that we don't
have multiple controllers updating the same resource multiple times, but
rely on the workplacement controller itself to update its status.

furthermore, there's a bug with the misschedule labels that make this
even more likely to occur.

it's not a silverbullet (as its in memory, if the controller dies before
it has the chance to reconcile, it's gone forever) but it will massively
reduce the chances of us losing the version forever