fix: system tests, and make test comparisons deterministic in unit tests
To fix the system tests, we added the 5-character hash to the end of the
pipeline-scoped user permission role. We needed this because resource
workflows also create a role with the same name otherwise, which was
failing the system tests by not creating the default resource workflow
permissions.

We also made the test comparisons deterministic in the unit tests by
using MatchFields instead of relying on the order.

Co-authored-by: Chunyi Lyu <chunyineko@gmail.com>