fix: remove skipping resource logic in workflow reconciler
This logic is a remnant of a time when the resource names had a random
UUID appended to them. The names are now completely deterministic, so we
can just do a create or update in the normal way in applyResources.

The tests have also been updated to no longer use the Generation to
assert that the resources haven't been changed. This is because the
fake client will reset the Generation in the Update call in applyResources.

We tried to use the ResourceVersion, but this gets incremented by the
fake client on every fetch, so it's also not useful.

We have manually verified that the resources have the same ResourceVersion
in a real kind cluster, and will add coverage in the system tests for
this behaviour.