feat: cross-namespace pipeline permissions (#210)
* fix: use labels to identify pipeline roles and rolebindings

Co-authored-by: Rich Barton-Cooper <rich@syntasso.io>

* feat: cross-namespace support in user pipeline permissions

Co-authored-by: Chunyi Lyu <chunyi@syntasso.io>
Co-authored-by: Sapphire Mason-Brown <sapphire@syntasso.io>

* fix: allow pipeline to create resources in other namespaces

Notable changes:
- Use a deterministic (but truncated) name for the cluster-wide resources.
  This allows us to track existing resources easier across time rather than
  UUIDs which were changing during code execution.
- Concatenate rules into ClusterRoles, by the namespace the ClusterRole
  is intended to give permissions in (via a namespaced RoleBinding).
  Previously we were only storing the last one listed of a certain group.
- Use the client.Object's type rather than GVK when comparing objects.
  The object's GVK is not populated consistently, so was sometimes nil.
  This meant that two objects with the same name/namespace were being
  treated as equal.
- Add more comprehensive tests for the RBAC logic. This helped us to
  narrow down the issues above.

* fix: system tests, and make test comparisons deterministic in unit tests

To fix the system tests, we added the 5-character hash to the end of the
pipeline-scoped user permission role. We needed this because resource
workflows also create a role with the same name otherwise, which was
failing the system tests by not creating the default resource workflow
permissions.

We also made the test comparisons deterministic in the unit tests by
using MatchFields instead of relying on the order.

Co-authored-by: Chunyi Lyu <chunyineko@gmail.com>

* fix: populate TypeMeta and use GVK in comparisons, and extract labels

Co-authored-by: Chunyi Lyu <chunyineko@gmail.com>

* fix: remove skipping resource logic in workflow reconciler

This logic is a remnant of a time when the resource names had a random
UUID appended to them. The names are now completely deterministic, so we
can just do a create or update in the normal way in applyResources.

The tests have also been updated to no longer use the Generation to
assert that the resources haven't been changed. This is because the
fake client will reset the Generation in the Update call in applyResources.

We tried to use the ResourceVersion, but this gets incremented by the
fake client on every fetch, so it's also not useful.

We have manually verified that the resources have the same ResourceVersion
in a real kind cluster, and will add coverage in the system tests for
this behaviour.

* chore: add tests for deterministic name generation

* chore: verify user permission role not recreated after a promise update

---------

Co-authored-by: Rich Barton-Cooper <rich@syntasso.io>
Co-authored-by: Sapphire Mason-Brown <sapphire@syntasso.io>