fix: apply bash-promise for all envs
We have introduced the ability to skip local/kind based set up
files when running system tests. There was one additional file
that was impacted and not making sure this is present on the LRE
caused failures in the soak testing.

Making sure that one file is installed, but still allowing us to
skip the local config files using the 'PLATFORM_SKIP_SETUP' var.