fix(#258): !improve status condition to reflect multiple pipelines
This commit:
- changes the `PipelineCompleted` condition type to
  `ConfigureWorkflowCompleted`
- updates the condition message to reflect the completion of multiple
  pipelines
- only sets the condition to `True` when the last pipeline in a
  Configure workflow has completed successfully

Previously, we set the `PipelineCompleted` condition to `True` at the
end of every successful Pipeline in a Configure workflow. This meant
there was a window in which the condition could be set to `True` before
all Pipelines had completed (before the next Pipeline was triggered and
the condition was reset to `False`).

This commit includes a change to `update-status` to only set the
`ConfigureWorkflowCompleted` condition to `True` when the last Pipeline
in the Configure workflow has completed successfully.

BREAKING CHANGE: Any automation relying on the `PipelineCompleted`
condition type will need to be updated to watch for the new
`ConfigureWorkflowCompleted` condition type instead. This is
semantically interchangeable, but reflects the completion of multiple
pipelines in a Configure workflow.

Co-authored-by: Jake Klein <jake@syntasso.io>