chore: improve core system tests
with the recent change to use a single worker as two different
destinations, the tests were no longer covering multi-destination
scheduling.

To validate that Kratix is indeed scheduling the workloads to the
appropriate place, we no longer can rely solely on the workload showing
up on the worker. Instead, we must check the WorkPlacements and verify
that the target destination for that workload is indeed the right
destination.

also removed a few duplicated assertions.