fix: ! use workflow- prefix in workflow labels
This commit replaces the work- prefix in workflow labels with a
workflow- prefix. The work- prefix was intended to solely be used for
Works.

BREAKING CHANGE: The labels are changing for workflow Jobs/Pods, which
means that when the running Kratix is upgraded to this version, we may
not be able to fetch old Jobs in the workflow reconciler. This might
result in the workflow being started again from the beginning. Users may
also have automation set up around the old labels, which will need to be
updated.

Co-authored-by: Alparslan Avci <alparslan@syntasso.io>