chore: consolidate on a single YAML parser
Both "gopkg.in/yaml.v2" and "sigs.k8s.io/yaml" were direct
dependencies. It's easier to reason about a single YAML parser,
and fewer dependencies is generally a good thing. So this PR
drops direct usage of the "gopkg.in/yaml.v2" parser.

But, this introduces a slight difference of behavior. A key differentiator
of the sigs.k8s.io/yaml parser is that it respects struct `json:""` tags
in addition to `yaml:""` tags. The "gopkg.in/yaml.v2" parser does not
do this. The result is that some MatchLabels struct fields now
serialize/deserialize as "matchLabels" rather than the previous
"matchlabels". This is likely a more "correct" behavior because it
matches the documentation, but any change has the potential to break
someone.