fix(#525): stop writing a test file to check perms
instead, rely on specific writers implementation.

For the GitWriter, we can do an empty push. If the credentials are
valid, Git should error with a "Already up-to-date" error. If they are
not valid, it will error with "Unauthorized" errors. There may be a
small race condition where the empty push fails but the repository is
not up-to-date, but that should be fixed in the next reconciliation.

For S3, we can initiate a multipart upload. If the server accepts the
initiation, then the credentials are valid. If not, an error is raised.
We can then abort the multipart upload.

Current installations of Kratix will leave the probe behind. Those would
need to be manually removed.

An improvement to this commit is to implement some logic to _not_ check
again on the next reconciliation of the State Store. Since the original
issue was on the excessive commits to the probe file, we can prioritise
this performance improvement later.

Co-authored-by: Chunyi Lyu <chunyi@syntasso.io>