fix: better handling of VersionID cache
Once the files are written successfully to the state store, we
immediately update a condition on the resource to reflect that. In rare
occasions, that status update can fail. When that happens, the version
id could be lost, since we were not storing it anywhere for the next
reconciliation

This commit fixes that behaviour. The Version ID is stored in the cache
immediately after it becomes available. It is only removed from the
cache when it's successfully persisted in the resource status.